+++ plone/app/event/recurrence.py
@@ -103,7 +103,7 @@ class EventOccurrenceAccessor(object):
     def __init__(self, context):
         object.__setattr__(self, 'context', context)
 
-        own_attr = ['start', 'end']
+        own_attr = ['start', 'end', 'url', 'parent']
         object.__setattr__(self, '_own_attr', own_attr)
 
     def _get_context(self, name):
@@ -122,3 +122,11 @@ class EventOccurrenceAccessor(object):
 
     def __delattr__(self, name):
         delattr(self._get_context(name), name)
+
+    # R/O properties
+    # TODO: Having uid here makes probably no sense, since Occurrences are
+    #       created on the fly and not persistent.
+
+    @property
+    def url(self):
+        return self.context.absolute_url()